Dev charlie1

+++ PolygonChecker/main.h
@@ -1,4 +1,4 @@
 #pragma once
 void printWelcome();
 int printShapeMenu();
-int* getTriangleSides(int* triangleSides);
+int* getTriangleSides(float* triangleSides);
